+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/info/GetSamplesInfo.groovy
@@ -37,6 +37,8 @@ class GetSamplesInfo implements GetSamplesInfoInput {
    */
   @Override
   void requestSampleInfosFor(String projectCode, Status status) {
+    Objects.requireNonNull(status, "Tried to request sample infos without providing a status.")
+    Objects.requireNonNull(projectCode, "Tried to request sample infos without providing a projectCode.")
     try {
         def sampleCodes = samplesDataSource.fetchSampleCodesFor(projectCode, status)
         def sampleCodesToNames = infoDataSource.fetchSampleNamesFor(sampleCodes)